move docs components into seperate files
