+++ mosaic-site/src/components/getting-started/design-cards/DesignCards.tsx
@@ -1,9 +1,6 @@
-import { Children, ReactNode } from "react";
-import { Image } from "@jpmorganchase/mosaic-site-components";
+import { Children } from "react";
 import useSplitArray from "../../../utils/useSplitArray";
 import styles from "./DesignCards.module.css";
-
-type DesignCardProps = { src: string; alt: string; children: ReactNode };
 
 export const DesignCards = ({ children }) => {
   const childrenArray = Children.toArray(children);
@@ -16,10 +13,3 @@ export const DesignCards = ({ children }) => {
     </div>
   );
 };
-
-export const DesignCard = ({ src, alt, children }: DesignCardProps) => (
-  <div className={styles.card}>
-    <Image src={src} alt={alt} nextImageClassName={""} />
-    <div className={styles.textContainer}>{children}</div>
-  </div>
-);
