+++ mosaic-site/src/components/getting-started/design-cards/index.ts
@@ -1,1 +1,2 @@
+export * from "./DesignCard";
 export * from "./DesignCards";
