+++ mosaic-site/src/components/getting-started/icons-table/IconsTable.tsx
@@ -1,7 +1,4 @@
 import { ReactElement, cloneElement, useRef, useEffect } from "react";
-import { SaltProvider, Density } from "@salt-ds/core";
-import { PrintIcon } from "@salt-ds/icons";
-
 import styles from "./IconsTable.module.css";
 
 type IconsTableProps = { children: ReactElement; caption: string };
@@ -22,18 +19,3 @@ export const IconsTable = ({ children, caption }: IconsTableProps) => {
     </div>
   );
 };
-
-type IconProps = {
-  density: Density;
-  iconSize: number;
-  children: ReactElement;
-};
-
-export const DensityIcon = ({ density, iconSize, children }: IconProps) => (
-  <>
-    <SaltProvider density={density}>
-      <PrintIcon size={iconSize} />
-    </SaltProvider>
-    <p className={styles.colored}>{children}</p>
-  </>
-);
