+++ mosaic-site/src/components/getting-started/icons-table/index.ts
@@ -1,1 +1,2 @@
+export * from "./DensityIcon";
 export * from "./IconsTable";
